main code
The design of this software is based on a task when I was working as a teaching assistant: to score the answers of two polling questions in each tutorials. The rule for giving points is that as long as one question is answered correctly, 1 point is awarded, all the wrong answers get 0.75, and no answers get 0 points. Because there are a total of 9 tutorials every week, a total of nearly 400 people need to be counted each time, so this automatic scoring system is designed.